feat: small enhancements and files location
